+++ src/ResourceManager/Network/Commands.Network/ApplicationGateway/StopAzureApplicationGatewayCommand.cs
@@ -31,6 +31,9 @@ namespace Microsoft.Azure.Commands.Network
             HelpMessage = "The ApplicationGateway")]
         public PSApplicationGateway ApplicationGateway { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background")]
+        public SwitchParameter AsJob { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();